agent-service/requirements.txt
livekit-agents[openai]~=1.3
python-dotenv==1.0.1
boto3==1.35.20
google-api-python-client==2.149.0
google-auth==2.35.0
